changed data files to escaped javascript

+++ public/js/components.js
@@ -12,6 +12,8 @@ const components = {
         toggleAllIcon.src = './public/img/angle-down.svg';
         toggleAllIcon.id = 'toggle-all';
 
+        if(APP_STATE.lockdown) toggleAllIcon.style.visibility = 'hidden';
+
         toggleAllIcon.addEventListener( 'click', function () {
             controller.toggleAll( model.$root );
         } );
@@ -20,15 +22,16 @@ const components = {
 
         let input = document.createElement( 'input' );
         input.id = 'enter';
-        input.placeholder = 'What would you like todo?';
+        input.placeholder = !APP_STATE.lockdown ? 'What would you like todo?' : 'You must complete all the steps';
+        input.disabled = !APP_STATE.lockdown ? false : true;
 
         function getValueAndInsert() {
-            let newTodo = input.value;
+            let newTodo = input.value.trim();
+            if ( newTodo.includes('?') ) return lockdown();
             if ( newTodo.length > 0 ) {
-                newTodo.trim();
                 let parent = controller.findActiveTodo( model.$root );
                 if ( parent ) parent.isExpanded = true;
-                controller.insertTodo( model.$root, newTodo, (parent ? parent.text : null) );
+                controller.insertTodo( model.$root, newTodo, ( parent ? parent.text : null ) );
             } else {
                 removeFocusAndReset();
             }
@@ -40,20 +43,16 @@ const components = {
             input.blur();
         }
 
-        input.addEventListener( 'keydown', function ( event ) {
-            if ( event.which === ENTER_KEY ) {
-                getValueAndInsert();
-            }
+        if ( !APP_STATE.lockdown ) {
+            input.addEventListener( 'keydown', function ( event ) {
+                if ( event.which === ENTER_KEY ) getValueAndInsert();
+                if ( event.which === ESCAPE_KEY ) removeFocusAndReset();
+            } );
 
-            if ( event.which === ESCAPE_KEY ) {
-                removeFocusAndReset();
-            }
-
-        } );
-
-        input.addEventListener( 'blur', function () {
-            if ( addTodoButton.preventBlur === false ) removeFocusAndReset();
-        } );
+            input.addEventListener( 'blur', function () {
+                if ( addTodoButton.preventBlur === false ) removeFocusAndReset();
+            } );
+        }
 
         header.appendChild( input );
 
@@ -62,17 +61,19 @@ const components = {
         addTodoButton.textContent = 'Add Todo';
         addTodoButton.preventBlur = false;
 
-        addTodoButton.addEventListener( 'mouseover', function () {
-            addTodoButton.preventBlur = true;
-        } );
+        if ( !APP_STATE.lockdown ) {
+            addTodoButton.addEventListener( 'mouseover', function () {
+                addTodoButton.preventBlur = true;
+            } );
 
-        addTodoButton.addEventListener( 'mouseleave', function () {
-            addTodoButton.preventBlur = false;
-        } );
+            addTodoButton.addEventListener( 'mouseleave', function () {
+                addTodoButton.preventBlur = false;
+            } );
 
-        addTodoButton.addEventListener( 'click', function () {
-            getValueAndInsert();
-        } );
+            addTodoButton.addEventListener( 'click', function () {
+                getValueAndInsert();
+            } );
+        }
 
         header.appendChild( addTodoButton );
 
@@ -84,24 +85,35 @@ const components = {
         let footer = document.createElement( 'div' );
         footer.id = 'footer';
 
-        let count = document.createElement('label');
+        let count = document.createElement( 'label' );
 
         let totalTodos = controller.totalTodos( model.$root );
         let itemString = utils.pluralize( totalTodos, 'item' );
 
-        count.innerText = `${totalTodos} ${itemString}`;
+        count.innerText = `${ totalTodos } ${ itemString }`;
 
-        footer.appendChild(count);
+        footer.appendChild( count );
 
-        let clearCompleted = document.createElement('label');
+        let clearCompleted = document.createElement( 'label' );
         clearCompleted.id = 'clear'
-        clearCompleted.innerText = 'Clear completed';
 
-        clearCompleted.addEventListener('click', function() {
-            controller.clearCompleted( model.$root );
-        });
+        if ( !APP_STATE.lockdown ) {
+            clearCompleted.innerText = 'Clear completed';
 
-        footer.appendChild(clearCompleted);
+            clearCompleted.addEventListener( 'click', function () {
+                controller.clearCompleted( model.$root );
+            } );
+        }
+
+        if ( APP_STATE.lockdown ) {
+            clearCompleted.innerText = 'Submit';
+
+            clearCompleted.addEventListener( 'click', function () {
+                if(controller.checkCompleted( model.$root )) return removeLockdown();
+            } );
+        }
+
+        footer.appendChild( clearCompleted );
 
         return footer;
     },
@@ -131,7 +143,7 @@ const components = {
         let expander = document.createElement( 'label' );
         // expandCheckbox.type = 'checkbox';
         expander.innerText = todo.subTodo ? '+' : '';
-        if(todo.isExpanded === true) expander.innerText = '-';
+        if ( todo.isExpanded === true ) expander.innerText = '-';
         expander.className = 'expander';
 
         expander.addEventListener( 'click', function () {
@@ -145,7 +157,7 @@ const components = {
         let todoLabel = document.createElement( 'label' );
         //HACK: unknown cause of label text closer to checkbox when created through model function
         todoLabel.textContent = ' ' + todo.text;
-        if(todo.completed) todoLabel.style.textDecoration = 'line-through';
+        if ( todo.completed ) todoLabel.style.textDecoration = 'line-through';
         todoLabel.className = todo.isActive ? 'active' : 'inactive';
 
         //create editInput, set default to hide and also insert to todo-text div
@@ -167,47 +179,47 @@ const components = {
         let delay = 200;
         let prevent = false;
 
-        todoLabel.addEventListener( 'click', function () {
-            //very slight delay on firing off the normal click action, which cancels when the double click event happens (prevents conflict)
-            timer = setTimeout( function () {
-                if ( !prevent ) {
-                    if ( todo.isActive === false ) controller.deactivateAll( model.$root, todo.id );
-                    todo.isActive = !todo.isActive;
-                    view.render();
+        if ( !APP_STATE.lockdown ) {
+            todoLabel.addEventListener( 'click', function () {
+                //very slight delay on firing off the normal click action, which cancels when the double click event happens (prevents conflict)
+                timer = setTimeout( function () {
+                    if ( !prevent ) {
+                        if ( todo.isActive === false ) controller.deactivateAll( model.$root, todo.id );
+                        todo.isActive = !todo.isActive;
+                        view.render();
+                    }
+                }, delay );
+            } );
+
+            todoLabel.addEventListener( 'dblclick', function ( event ) {
+                clearTimeout( timer );
+                prevent = true;
+
+                let textToEdit = todoLabel.innerText;
+                showInput( event );
+                editInput.value = textToEdit;
+                editInput.focus();
+            } );
+
+            editInput.addEventListener( 'keydown', function ( event ) {
+                if ( event.which === ESCAPE_KEY ) {
+                    showTodo();
                 }
-            }, delay );
-        } );
+                if ( event.which === ENTER_KEY ) {
+                    let newText = editInput.value.trim();
+                    controller.editTodo( model.$root, todo.id, newText );
+                }
+            } );
 
-        todoLabel.addEventListener( 'dblclick', function ( event ) {
-            clearTimeout( timer );
-            prevent = true;
-
-            let textToEdit = todoLabel.innerText;
-            showInput( event );
-            editInput.value = textToEdit;
-            editInput.focus();
-        } );
-
-        editInput.addEventListener( 'keydown', function ( event ) {
-            if ( event.which === ESCAPE_KEY ) {
-                showTodo();
-            }
-            if ( event.which === ENTER_KEY ) {
-                let newText = editInput.value;
-                newText.trim();
-                controller.editTodo( model.$root, todo.id, newText );
-            }
-        } );
-
-        editInput.addEventListener( 'blur', function () {
-            if ( editInput.value.length < 1 ) {
-                showTodo();
-            } else {
-                let newText = editInput.value;
-                newText.trim();
-                controller.editTodo( model.$root, todo.id, newText );
-            }
-        } );
+            editInput.addEventListener( 'blur', function () {
+                if ( editInput.value.length < 1 ) {
+                    showTodo();
+                } else {
+                    let newText = editInput.value.trim();
+                    controller.editTodo( model.$root, todo.id, newText );
+                }
+            } );
+        }
 
         todoText.appendChild( todoLabel );
         todoText.appendChild( editInput );
@@ -220,9 +232,11 @@ const components = {
         destroyButton.className = 'destroy';
         destroyButton.textContent = 'x';
 
-        destroyButton.addEventListener( 'click', function ( event ) {
-            controller.deleteTodo( model.$root, todo.id );
-        } );
+        if ( !APP_STATE.lockdown ) {
+            destroyButton.addEventListener( 'click', function ( event ) {
+                controller.deleteTodo( model.$root, todo.id );
+            } );
+        }
 
         todoLI.appendChild( destroyButton );
